requirements.txt
matplotlib==3.3.4
mpi4py==3.0.3
numpy==1.20.1
pandas==1.2.3
Pillow==9.2.0
scikit_learn==1.1.2
seaborn==0.11.1
setuptools==45.2.0
torch==1.8.0+cu111
torchvision==0.9.0+cu111
